Remove measure v1
