+++ utils/measure/Dockerfile
