--- utils/measure_v2/.gitignore
@@ -1,5 +1,0 @@
-.env
-export/
-measure
-light_controller/.python_hue
-measure.log
