--- utils/measure_v2/requirements.txt
@@ -1,25 +1,0 @@
-anyio==3.3.0
-asyncclick==7.1.2.3
-certifi==2021.5.30
-charset-normalizer==2.0.4
-gensim==3.8.1
-idna==3.2
-importlib-metadata==4.8.1
-phue==1.1
-prompt-toolkit==1.0.14
-pycryptodome==3.10.1
-Pygments==2.10.0
-PyInquirer==1.0.3
-python-dateutil==2.8.2
-python-decouple==3.4
-python-kasa==0.4.0.dev3
-regex==2021.8.28
-requests==2.26.0
-six==1.16.0
-smart-open==5.2.1
-sniffio==1.2.0
-tinytuya==1.2.7
-tuyapower==0.0.25
-urllib3==1.26.6
-wcwidth==0.2.5
-zipp==3.5.0
